Don't overwrite existing posts

+++ statirator/blog/management/commands/create_post.py
@@ -38,10 +38,13 @@ class Command(BaseCommand):
         }
 
         filename = os.path.join(get_blog_dir(), slug + '.rst')
+
+        if os.path.exists(filename):
+            raise CommandError('"{0}" exists, aborting'.format(filename))
         with codecs.open(filename, 'w', 'utf-8') as post_file:
 
-            metadata_rendered = render_to_string('blog/new_post_metadata.txt', ctx)
-            post_file.write(metadata_rendered)
+            meta_rendered = render_to_string('blog/new_post_metadata.txt', ctx)
+            post_file.write(meta_rendered)
 
             cur_lang = translation.get_language()
 
@@ -51,7 +54,7 @@ class Command(BaseCommand):
                 translation.activate(lang_code)
 
                 ctx.update({'title': title_or_slug})
-                post_rendered = render_to_string('blog/new_post_content.txt', ctx)
-                post_file.write(post_rendered)
+                rendered = render_to_string('blog/new_post_content.txt', ctx)
+                post_file.write(rendered)
 
             translation.activate(cur_lang)
